% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{split_df}
\alias{split_df}
\title{Split a large dataframe into multiple smaller data frames.}
\usage{
split_df(data, n)
}
\description{
\code{split_df} splits a dataframe into n (nearly) equal pieces, all 
pieces containing all columns of the original data frame. If the number 
of rows in the original dataframe is not evenly divisibile by n, the nth
dataframe will contain the remainder rows.
}
\details{
@param data Data frame. 
 @param n Integer. Number of divisions.
 @return Returns a list of dataframes. All dataframes in list will 
         comprise the original data frame.
}

